# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shadow Relay API",
    "version": "0.2.0",
    "description": "私有源控制面。管理接口使用 Bearer 管理令牌，订阅令牌仅授权相应编排组及格式。完整语义见 docs/。"
  },
  "security": [
    {
      "adminBearer": []
    }
  ],
  "components": {
    "securitySchemes": {
      "adminBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
    "/api/v1/sources": {
      "get": {
        "summary": "读取 sources",
        "operationId": "get_api_v1_sources",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Not found or not authorized"
          },
          "409": {
            "description": "State conflict"
          }
        }
      }
    },
    "/api/v1/catalogs": {
      "get": {
        "summary": "读取 catalogs",
        "operationId": "get_api_v1_catalogs",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Not found or not authorized"
          },
          "409": {
            "description": "State conflict"
          }
        }
      },
      "post": {
        "summary": "创建 catalogs",
        "operationId": "post_api_v1_catalogs",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Not found or not authorized"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/source-sets": {
      "get": {
        "summary": "读取 source-sets",
        "operationId": "get_api_v1_source_sets",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Not found or not authorized"
          },
          "409": {
            "description": "State conflict"
          }
        }
      },
      "post": {
        "summary": "创建 source-sets",
        "operationId": "post_api_v1_source_sets",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Not found or not authorized"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceSetInput"
              }
            }
          }
        }
      }
    },
    "/api/v1/publications": {
      "get": {
        "summary": "读取 publications",
        "operationId": "get_api_v1_publications",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Not found or not authorized"
          },
          "409": {
            "description": "State conflict"
          }
        }
      }
    },
    "/api/v1/bindings": {
      "get": {
        "summary": "读取 bindings",
        "operationId": "get_api_v1_bindings",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Not found or not authorized"
          },
          "409": {
            "description": "State conflict"
          }
        }
      },
      "post": {
        "summary": "创建客户端绑定并只返回一次令牌",
        "operationId": "post_api_v1_bindings",
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Not found or not authorized"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/v1/runtimes": {
      "get": {
        "summary": "读取 runtimes",
        "operationId": "get_api_v1_runtimes",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Not found or not authorized"
          },
          "409": {
            "description": "State conflict"
          }
        }
      },
      "post": {
        "summary": "创建 runtimes",
        "operationId": "post_api_v1_runtimes",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Not found or not authorized"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jobs": {
      "get": {
        "summary": "读取 jobs",
        "operationId": "get_api_v1_jobs",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Not found or not authorized"
          },
          "409": {
            "description": "State conflict"
          }
        }
      }
    },
...
    "/api/v1/adapters": {
      "get": {
        "summary": "读取 adapters",
        "operationId": "get_api_v1_adapters",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Not found or not authorized"
          },
          "409": {
            "description": "State conflict"
          }
        }
      }
    },
...
    "/api/v1/sources/{id}": {
      "get": {
        "summary": "读取源",
        "operationId": "get_api_v1_sources_id",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Not found or not authorized"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "put": {
        "summary": "编辑源元信息（端点与协议不可变）",
        "operationId": "put_api_v1_sources_id",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Not found or not authorized"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "additionalProperties": false,
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "url": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "protocol": {
                    "type": "string"
                  },
                  "network": {
                    "type": "string"
                  },
                  "trust": {
                    "type": "string"
                  },
                  "mode": {
                    "type": "string"
                  },
                  "runtimeId": {
                    "type": "string"
                  },
                  "updatePolicy": {
                    "type": "string"
                  },
                  "intervalMinutes": {
                    "type": "integer"
                  },
                  "mediaTypes": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "headers": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "删除无编排引用的源",
        "operationId": "delete_api_v1_sources_id",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Not found or not authorized"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sources/{id}/revisions": {
      "get": {
        "summary": "源 revisions",
        "operationId": "get_api_v1_sources_id_revisions",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Not found or not authorized"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/sources/{id}/probes": {
      "get": {
        "summary": "源 probes",
        "operationId": "get_api_v1_sources_id_probes",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Not found or not authorized"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/sources/{id}/revisions/{revision}/diff": {
      "get": {
        "summary": "版本 diff",
        "operationId": "get_api_v1_sources_id_revisions_revision_diff",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Not found or not authorized"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "revision",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/sources/{id}/revisions/{revision}/raw": {
      "get": {
        "summary": "版本 raw",
        "operationId": "get_api_v1_sources_id_revisions_revision_raw",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Not found or not authorized"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "revision",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/catalogs/{id}": {
      "put": {
        "summary": "编辑 catalogs",
        "operationId": "put_api_v1_catalogs_id",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Not found or not authorized"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
...
    "/api/v1/publications/{id}": {
      "get": {
        "summary": "读取不可变发布",
        "operationId": "get_api_v1_publications_id",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation failed"
          },
          "404": {
            "description": "Not found or not authorized"
          },
          "409": {
            "description": "State conflict"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  }
}